added claims and enhance the build/test cycle (#3)
* added claims and enhance the build/test cycle

This patch updates the oidcmock primarily to have extra claims.

However, I was unable to build it since it could not pull the missing
package 'github.com/aporeto-inc/regodux/versions' so created a script
mkversions.sh and modified the Makefile to build the new versions.go
file when needed.

During the course of testing, it was useful to push sha tagged docker
images instead of always pushing the `latest` image, which will break
the working FT tests in the CI pipeline

So this Makefile pushes the docker image tagged as with git branch
tag and the 8 chars of the SHA hash. It will only push tag `latest`
when building on master.

This means that after this PR is merged (or any PR) you must checkout
the `master` branch and build and push it before the FT tests can use
it (unless there is a CI already in place for this ?? )

It also tags the image and version string with `-dirty` if there are
uncommitted changes during the build.

I also made it build a local version of the binary `oidcmock` for testing,
and only builds the 386 architecture for the docker image. This lets you
build and test locally on non-Linux OS's (i.e. OS X)

[543] $ uname -s
Darwin

[544] $ ./oidcmock --version
v0.12 - update-oidc-mock (e8cbd29b-dirty)

Also fixed the banner to display the correct version, the same as `oidcmock --version`

◼︎ OIDC-MOCK  v0.12 - update-oidc-mock (e8cbd29b-dirty)

* add SRC to makefile and add oidc claims

* remove more stuff on make clean, fixup docker push

* added VERSION_FILE to clean target

* fixed a makefile variable

* fixed makefile targets

* updated README

* updated Makefile clean target

* changed array to slice

* Makefile: fixed clean target to remove docker/oidcmock

* Makefile: cleanup targets

* Makefile: push to GCR with -dirty tag if the working dir is dirty

* Makefile: echo when docker tag is successful